Modified Model and DAOs. Update Servlet to create alarms.

+++ teleassistence-app/src/es/isst/ca/dao/EventDAO.java
@@ -7,7 +7,7 @@ import es.isst.ca.model.Location;
 
 public interface EventDAO {
 
-	public void addLocation(String originator, Long timestamp, Long latitude, Long longitude);
+	public void addLocation(String originator, Long timestamp, Double latitude, Double longitude);
 	public void addAcceleration(String originator, Long timestamp, Double acceleration_x_axis, Double acceleration_y_axis, Double acceleration_z_axis);
 	
 	public List<Event> listEvents(String originator);
